React Crypto App
I was feeling particularly motivated today so i have gone ahead and coded up more or less the entire base for this application. I wanted to practice more with React Router and needed a project to go along with it.

For the most part it is just fetching and showing data from a couple of coin geckos api's. I have used react-router-dom to allow the user to click a crypto from the top 10 based on coin gecko rank and allow them to pull up some data about it. This was a tedious process thanks to some of coin gecko's chosen variable names but the functionality is all there now.

After clicking a coin you will be taken to a new route showing the user more in depth info and statistics they may want to see.

I will add some search functionality as well as a means of setting the amount of search results which i don't think will be too hard soon as well as some other features i have in mind but i am not certain of yet.