Search and increase results functionality
Added: Search bar - This was very easy to implement thanks to our navigation being handled through the url so it is as simple as navigating to the url with the query appended. This is obviously not a perfect system and as of yet does not allow for the searching of its shortform name i.e. BTC it also does not have an autocomplete feature, yet.

Added: Load more button - i have added a button to the bottom of the page that allows a user to get more than 10 results this is limited at 100 results due to me using the free version of the API package however in the real world this would continue on until the last crypto the API is capable of churning back. or then be split into multiple pages.

Changed: Styles - I have altered some styles to be more crisp and clean and easy to maintain or customise.